# optimize AddMod where modulus is 4-limbs wide and x/y highest limb is  less than or equal to modulus highest limb