uint256 master:
```
BenchmarkAddMod/small/uint256-4                 71864941                14.78 ns/op
BenchmarkAddMod/mod64/uint256-4                 45851817                23.58 ns/op
BenchmarkAddMod/mod128/uint256-4                23512131                49.20 ns/op
BenchmarkAddMod/mod192/uint256-4                22141264                53.37 ns/op
BenchmarkAddMod/mod256/uint256-4                55604551                21.38 ns/op
```

this PR:
```
BenchmarkAddMod/small/uint256-4                 74203911                14.75 ns/op
BenchmarkAddMod/mod64/uint256-4                 44762869                23.61 ns/op
BenchmarkAddMod/mod128/uint256-4                22504081                49.59 ns/op
BenchmarkAddMod/mod192/uint256-4                21490502                52.70 ns/op
BenchmarkAddMod/mod256/uint256-4                65233677                17.39 ns/op
```

benchmarks ran with turbo-boost disabled so there is little deviation in the results between runs.